add some checks to setting up localtime

+++ library/pac-build.build.sh
@@ -45,8 +45,9 @@ create_chroot() {
 	fi
 
 	install -m644 "${pacman_conf_path}" "${builddir}/etc/pacman.conf"
-
-	ln -s "/usr/share/zoneinfo/${timezone}" ${builddir}/etc/localtime
+	if [ ! -L ${builddir}/etc/localtime ] || [ ! -f ${builddir}/etc/localtime ]; then
+		ln -s "/usr/share/zoneinfo/${timezone}" ${builddir}/etc/localtime
+	fi
 }
 
 check_mountfs() {
